cmake: fix the build on aarch64
x86 platforms generally use a default-signed char, but !x86 does
not necessarily.  Force it with -fsiged-char so that all platforms
assume the correct signedness, which fixes the build on aarch64 at
the very least.

This is a nop on x86, so we don't bother trying to scope it beyond
GCC/Clang.