Implement global asset cache
This cache is intended to load and unload assets from a unified data
structure. This helps with moving all LoadFileInMem function (which is a
dynamic memory allocation function) into a well-defined sequence in the
code.

This change only introduces the global cache. It does not replace the
existing Diablo LoadFileInMem behavior.