Delete setmaps.h and setmaps.cpp
The files are related to generation of quest maps, which exist as
quest content that is optional to completing the Diablo quest line. This
is not useful for a map generation program with the goal of finding
optimal routing.

The function setmaps.cpp:ObjIndex(x,y) is moved to objects.cpp, as it is
the only file using this funcion now.